requirements.txt
appdirs==1.4.3
cached-property==1.3.1
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
dill==0.2.7.1
Django==1.11
django-braces==1.11.0
django-cleanup==1.0.1
django-filter==1.0.4
djangorestframework==3.6.2
djangorestframework-queryfields==1.0.0
enum34==1.1.6
fcm-django==0.2.13
firebase-admin==2.8.0
future==0.16.0
futures==3.2.0
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-core==0.28.0
google-cloud-firestore==0.28.0
google-cloud-storage==1.7.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpcio==1.8.4
idna==2.6
isodate==0.5.4
lxml==4.0.0
Markdown==2.6.6
oauthlib==2.0.1
olefile==0.44
osa==0.2
packaging==16.8
Pillow==4.1.1
ply==3.8
protobuf==3.5.1
psycopg2==2.7.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyfcm==1.4.3
pyparsing==2.2.0
pyreadline==2.1
pytz==2017.2
requests==2.18.4
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.10.0
urllib3==1.22
zeep==2.4.0
